requirements.txt
pillow
opencv-python
tensorflow
streamlit
keras
pillow